+++ .eslintrc.js
@@ -21,5 +21,7 @@ module.exports = {
     'import/prefer-default-export': 'off',
     'no-plusplus': 'off',
     'no-unused-expressions': 'off',
+    'func-names': 'off',
+    'no-console': 'off',
   },
 };
