+++ src/slider.test.js
@@ -1,6 +1,31 @@
-import { createSlider, getAllCarouselItems } from './slider';
+import { getAllCarouselItems, nextSlide } from './slider';
 import '@testing-library/jest-dom';
 
+function createSlider(element, countSliders = 1) {
+  element.classList.add('carousel');
+  element
+    .appendChild(document.createElement('button'))
+    .classList.add('button', 'preview');
+  element.appendChild(document.createElement('ul'));
+  element
+    .appendChild(document.createElement('button'))
+    .classList.add('button', 'next');
+
+  const ul = element.querySelector('ul');
+  for (let i = 0; i < countSliders; i++) {
+    const li = document.createElement('li');
+    const img = document.createElement('img');
+    img.classList.add('carousel-item');
+    if (i === 0) {
+      img.classList.add('active');
+    }
+    li.appendChild(img);
+    ul.appendChild(li);
+  }
+
+  element.querySelector('button.next').addEventListener('click', nextSlide);
+  element.querySelector('button.preview').addEventListener('click', nextSlide);
+}
 describe('Get all items', () => {
   let el;
 
@@ -21,7 +46,7 @@ describe('Get all items', () => {
   it('All carousel items is presented', () => {
     expect(getAllCarouselItems(el).length).toBe(3);
   });
-  it.skip("Button 'next' switches slide to the next", () => {
+  it("Button 'next' switches slide to the next", () => {
     el.querySelector('button.next').click();
     expect(getAllCarouselItems(el)[0].classList.contains('active')).toBeFalsy;
     expect(getAllCarouselItems(el)[1].classList.contains('active')).toBeTruthy;
